+++ src/app/locales/en/components/-browser-columns-name/filesystem.js
@@ -1,4 +1,5 @@
 export default {
+  files: 'Files',
   size: 'Size',
   owner: 'Owner',
   modification: 'Modification',