+++ src/app/locales/en/components/archive-browser/table-head-row.js
@@ -1,7 +1,5 @@
+import headers from '../-browser-columns-name/archive';
+
 export default {
-  archives: 'Archives',
-  creator: 'Creator',
-  state: 'State',
-  created: 'Created at',
-  incremental: 'Base archive',
+  headers,
 };